refactor(library/vue): moves "/definition.ts" to "/core.ts" in "/Ref"
